refactor(View): InputView, OutputView 인터페이스로 분리
