+++ src/main/java/ladder/Application.java
@@ -1,5 +1,7 @@
 package ladder;
 
+import ladder.config.AppConfig;
+
 public class Application {
     public static void main(String[] args) {
         AppConfig.ladderController().run();
